// src/listing/interfaces/item.interface.ts
export interface Item {
  defindex: number;
  quality: number;
  craftable: boolean;
  killstreak: number;
  australium: boolean;
  festive: boolean;
  effect: number;
  paintkit: number;
  wear: number;
  quality2: number;
  target: number;
  craftnumber: number;
  crateseries: number;
  output: number;
  outputQuality: number;
}